Revert CNAME to ayab-knitting.com
It looks like GitHub requires the DNS entry to actually point to `allyarnsarebeautiful.github.io` before it will generate a TLS certificate. Currently `www.ayab-knitting.com` points to `ayab-knitting.com` (which resolves to the same IP as `allyarnsarebeautiful.github.io` but that's not enough apparently).

Until this can be changed in DNS, let's revert.